+++ packages/react-network/src/server.ts
@@ -26,6 +26,16 @@ export function applyToContext<T extends Context>(
   Object.entries(cookies).forEach(([cookie, options]) => {
     const {value, ...cookieOptions} = options;
 
+    // only set cookies that have been changed
+    // decode URI as the manager's cookie value is also decoded
+    const rawCookieValue = ctx.cookies.get(cookie);
+    if (
+      rawCookieValue != null &&
+      decodeURIComponent(rawCookieValue) === value
+    ) {
+      return;
+    }
+
     // missing 'none` in `sameSite`
     // https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/cookies/index.d.ts#L91
     ctx.cookies.set(cookie, value, cookieOptions as any);